+++ app/Http/Controllers/Manager/OptionsController.php
@@ -24,6 +24,7 @@ declare(strict_types=1);
 namespace Adshares\Adserver\Http\Controllers\Manager;
 
 use Adshares\Adserver\Client\Mapper\AbstractFilterMapper;
+use Adshares\Adserver\Exceptions\MissingInitialConfigurationException;
 use Adshares\Adserver\Http\Controller;
 use Adshares\Adserver\Http\Requests\TargetingReachRequest;
 use Adshares\Adserver\Models\Config;
@@ -33,6 +34,7 @@ use Adshares\Adserver\ViewModel\OptionsSelector;
 use Adshares\Common\Application\Service\ConfigurationRepository;
 use Illuminate\Http\Request;
 use Symfony\Component\HttpFoundation\JsonResponse;
+use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 
 class OptionsController extends Controller
 {
@@ -82,13 +84,23 @@ class OptionsController extends Controller
 
     public function media(): JsonResponse
     {
-        return self::json($this->optionsRepository->fetchMedia()->toArray());
+        try {
+            $media = $this->optionsRepository->fetchMedia();
+        } catch (MissingInitialConfigurationException $exception) {
+            return self::json();
+        }
+        return self::json($media->toArray());
     }
 
     public function medium(string $medium, Request $request): JsonResponse
     {
         $vendor = $request->get('vendor');
-        $data = $this->optionsRepository->fetchMedium($medium, $vendor)->toArray();
+        try {
+            $mediumObject = $this->optionsRepository->fetchMedium($medium, $vendor);
+        } catch (MissingInitialConfigurationException $exception) {
+            throw new NotFoundHttpException($exception->getMessage());
+        }
+        $data = $mediumObject->toArray();
 
         if ($request->get('e')) {
             foreach ($data['targeting']['site'] ?? [] as $key => $value) {
@@ -106,7 +118,12 @@ class OptionsController extends Controller
     public function vendors(string $medium): JsonResponse
     {
         $data = [];
-        foreach ($this->optionsRepository->fetchTaxonomy()->getMedia() as $mediumObject) {
+        try {
+            $taxonomy = $this->optionsRepository->fetchTaxonomy();
+        } catch (MissingInitialConfigurationException $exception) {
+            return self::json();
+        }
+        foreach ($taxonomy->getMedia() as $mediumObject) {
             if ($mediumObject->getName() === $medium && $mediumObject->getVendor() !== null) {
                 $data[$mediumObject->getVendor()] = $mediumObject->getVendorLabel();
             }
@@ -122,7 +139,12 @@ class OptionsController extends Controller
                 '/site/quality' => true
             ];
         }
-        return self::json(new OptionsSelector($this->optionsRepository->fetchTargetingOptions()->exclude($exclusions)));
+        try {
+            $selector = $this->optionsRepository->fetchTargetingOptions();
+        } catch (MissingInitialConfigurationException $exception) {
+            return self::json();
+        }
+        return self::json(new OptionsSelector($selector->exclude($exclusions)));
     }
 
     public function targetingReach(TargetingReachRequest $request): JsonResponse
@@ -148,7 +170,12 @@ class OptionsController extends Controller
                 ) => true
             ];
         }
-        return self::json(new OptionsSelector($this->optionsRepository->fetchFilteringOptions()->exclude($exclusions)));
+        try {
+            $selector = $this->optionsRepository->fetchFilteringOptions();
+        } catch (MissingInitialConfigurationException $exception) {
+            return self::json();
+        }
+        return self::json(new OptionsSelector($selector->exclude($exclusions)));
     }
 
     public function languages(): JsonResponse