+++ CHANGELOG.md
@@ -8,6 +8,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Filtering local banners in site's classifier
+### Fixed
+- Error while targeting/filtering is not cached
+
+## [1.16.2] - 2022-04-21
+### Fixed
+- Admin dashboard reports
 
 ## [1.16.1] - 2022-04-07
 ### Added
@@ -606,7 +612,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.1.0]
 
-[Unreleased]: https://github.com/adshares/adserver/compare/v1.16.1...develop
+[Unreleased]: https://github.com/adshares/adserver/compare/v1.16.2...develop
+[1.16.2]: https://github.com/adshares/adserver/compare/v1.16.1...v1.16.2
 [1.16.1]: https://github.com/adshares/adserver/compare/v1.16.0...v1.16.1
 [1.16.0]: https://github.com/adshares/adserver/compare/v1.15.0...v1.16.0
 [1.15.0]: https://github.com/adshares/adserver/compare/v1.14.2...v1.15.0